# docs/swagger.yaml
basePath: /api
definitions:
  controllers.AssignRolePayload:
    properties:
      role_id:
        type: integer
    type: object
...
  controllers.ErrorResponseD:
    properties:
      error:
        type: string
    type: object
...
  controllers.PermissionDataRequest:
    properties:
      permission_id:
        example: 1
        type: integer
    type: object
  controllers.PermissionResponse:
    properties:
      active:
        example: true
        type: boolean
      description:
        example: Permiso de Crud Usuarios
        type: string
      id:
        example: 1
        type: integer
      module_id:
        example: 1
        type: integer
      name:
        example: Gestion Usuarios
        type: string
    type: object
...
  controllers.RoleResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  controllers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  controllers.UserRoleResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: seri-api-utn-2024.fly.dev
info:
  contact:
    email: jfpastillod@utn.edu.ec
    name: Pastillo D Joan
    url: https://www.utn.edu.ec
  description: Documentación de LA API DE SEGURIDAD hecha con Go.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: API SEGURIDAD
  version: "1.9"
paths:
...
  /permissions/all:
    get:
      description: Lista todos los permisos disponibles, incluyendo los módulos a
        los que pertenecen
      produces:
      - application/json
      responses:
        "200":
          description: Lista de permisos
          schema:
            items:
              $ref: '#/definitions/controllers.PermissionResponse'
            type: array
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todos los permisos
      tags:
      - Rol_Permisos
...
  /roles/{role_id}/permissions:
    delete:
      consumes:
      - application/json
      description: Elimina la asociación de un permiso específico con un rol
      parameters:
      - description: ID del rol
        in: path
        name: role_id
        required: true
        type: integer
      - description: Datos del permiso a eliminar
        in: body
        name: permissionData
        required: true
        schema:
          $ref: '#/definitions/controllers.PermissionDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar permiso de rol
      tags:
      - Rol_Permisos
    get:
      description: Lista todos los permisos asignados a un rol específico
      parameters:
      - description: ID del rol
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de permisos
          schema:
            items:
              $ref: '#/definitions/controllers.PermissionResponse'
            type: array
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener permisos de un rol
      tags:
      - Rol_Permisos
    post:
      consumes:
      - application/json
      description: Asocia un permiso existente a un rol específico
      parameters:
      - description: ID del rol
        in: path
        name: role_id
        required: true
        type: integer
      - description: Datos del permiso a asignar
        in: body
        name: permissionData
        required: true
        schema:
          $ref: '#/definitions/controllers.PermissionDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: role_permission
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Asignar permiso a rol
      tags:
      - Rol_Permisos
...
  /users/{id}/roles:
    get:
      description: Devuelve una lista de los roles asignados a un usuario especificado
        por su ID.
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de roles asignados
          schema:
            items:
              $ref: '#/definitions/controllers.RoleResponse'
            type: array
        "400":
          description: ID de usuario inválido
          schema:
            $ref: '#/definitions/controllers.ErrorResponseD'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/controllers.ErrorResponseD'
      security:
      - BearerAuth: []
      summary: Obtener roles de un usuario
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Asigna un rol a un usuario especificado por su ID.
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Estructura JSON con el ID del rol
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AssignRolePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Rol asignado correctamente
          schema:
            $ref: '#/definitions/controllers.UserRoleResponse'
        "400":
          description: ID de usuario o datos inválidos
          schema:
            $ref: '#/definitions/controllers.ErrorResponseD'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/controllers.ErrorResponseD'
      security:
      - BearerAuth: []
      summary: Asignar un rol a un usuario
      tags:
      - Roles
  /users/{id}/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Elimina un rol previamente asignado a un usuario dado su ID.
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: ID del rol a eliminar
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mensaje de éxito indicando que el rol fue eliminado
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: ID de usuario o rol inválido
          schema:
            $ref: '#/definitions/controllers.ErrorResponseD'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/controllers.ErrorResponseD'
      security:
      - BearerAuth: []
      summary: Eliminar un rol de un usuario
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"